+++ contents/winrmexe.rb
@@ -9,12 +9,24 @@ shell = ENV['RD_CONFIG_SHELL']
 realm = ENV['RD_CONFIG_KRB5_REALM']
 command = ENV['RD_EXEC_COMMAND']
 winrmtimeout = ENV['RD_CONFIG_WINRMTIMEOUT']
+override = ENV['RD_CONFIG_ALLOWOVERRIDE']
+host = ENV['RD_OPTION_WINRMHOST'] if ENV['RD_OPTION_WINRMHOST'] and (override == 'host' or override == 'all')
+user = ENV['RD_OPTION_WINRMUSER'] if ENV['RD_OPTION_WINRMUSER'] and (override == 'user' or override == 'all')
+pass = ENV['RD_OPTION_WINRMPASS'].dup if ENV['RD_OPTION_WINRMPASS'] and (override == 'user' or override == 'all')
+
 endpoint = "http://#{host}:#{port}/wsman"
 output = ''
 
+# Wrapper to fix: "not setting executing flags by rundeck for 2nd file in plugin"
+# # https://github.com/rundeck/rundeck/issues/1421
+# remove it after issue will be fixed
+if File.exist?("#{ENV['RD_PLUGIN_BASE']}/winrmcp.rb") and not File.executable?("#{ENV['RD_PLUGIN_BASE']}/winrmcp.rb")
+    File.chmod(0764, "#{ENV['RD_PLUGIN_BASE']}/winrmcp.rb")
+end
+
 # Wrapper ro avoid strange and undocumented behavior of rundeck
 # Should be deleted after rundeck fix
-#  ''"'"'
+# https://github.com/rundeck/rundeck/issues/602
 command = command.gsub(/'"'"'' /, '\'')
 command = command.gsub(/ ''"'"'/, '\'')
 command = command.gsub(/ '"/, '"')
@@ -25,11 +37,11 @@ command = command.gsub(/"' /, '"')
 # replace rm -f into rm -force
 # auto copying renames file from .sh into .ps1
 # so in that case we should call file with ps1 extension
+# TODO: add extension based on shell variable
+# TODO: deleting based on shell variable
 exit 0 if command.include? 'chmod +x /tmp/'
 command = command.gsub(%r{rm -f /tmp/}, 'rm -force /tmp/') if command.include? 'rm -f /tmp/'
 command = command.gsub(/\.sh/, '.ps1') if %r{/tmp/.*\.sh}.match(command)
-
-# TODO: ENV['WINRM_LOG'] = '' or 'debug, info, warn, or error'
 
 if ENV['RD_JOB_LOGLEVEL'] == 'DEBUG'
   puts 'variables is:'
@@ -37,15 +49,15 @@ if ENV['RD_JOB_LOGLEVEL'] == 'DEBUG'
   puts "endpoint is #{endpoint}"
   puts "user is #{user}"
   puts "pass is ********"
-  #  puts "pass is #{pass}"
+  #  puts "pass is #{pass}" # uncomment it for full auth debugging
   puts "command is #{ENV['RD_EXEC_COMMAND']}"
   puts "newcommand is #{command}"
 
   puts 'ENV'
   ENV.each do |k, v|
     puts "#{k} => #{v}" if v != pass
-    puts "#{k} => ********" if v == pass
-#    puts "#{k} => #{v}" if v == pass
+    puts "#{k} => ********" if v == pass or k == 'RD_CONFIG_PASS'
+#    puts "#{k} => #{v}" if v == pass # uncomment it for full auth debugging
   end
 end